moved payload files to differents folders
