+++ src/payload/index.ts
@@ -4,7 +4,7 @@ import payload from "payload";
 import { InitOptions } from "payload/config";
 
 dotenv.config({
-  path: path.resolve(__dirname, "../.env"),
+  path: path.resolve(__dirname, "../../.env"),
 });
 let cached = (global as any).payload;
 if (!cached) {
