+++ tsconfig.server.json
@@ -5,7 +5,8 @@
     "moduleResolution": "node",
     "outDir": "dist",
     "noEmit": false,
-    "jsx": "react"
+    "jsx": "react",
+    "allowImportingTsExtensions": true
   },
   "include": ["src/server.ts", "src/payload.config.ts"],
   "exclude": ["node_modules"]
